modularized graph portion
