+++ src/index.js
@@ -5,7 +5,13 @@ import { PlaceData } from "./api/PlaceData";
 import { ColorData } from "./api/ColorData";
 import { RouteData } from "./api/RouteData";
 import { TurfData } from "./api/TurfData";
+import { Graph } from "./api/Graph";
 
+/**
+ * ngraph.graph helper functions for weighted graph computation
+ * @property {Graph}
+ */
+export { Graph };
 /**
  * Turf.js data and geometry helpers
  * @property {TurfData}
